+++ test/issues.js
@@ -11,4 +11,5 @@ describe('Issues.', function () {
   require('./issues/issue-46.js');
   require('./issues/issue-54.js');
   require('./issues/issue-64.js');
+  require('./issues/issue-parse-function-security.js');
 });